MichalJabi changes request Resolved
