+++ week-1-ts-ecosystem/w01e01-typescript-basics/src/User.d.ts
@@ -1,0 +1,11 @@
+type Address = {
+    street: string,
+    no: number
+}
+
+export type UserType = {
+    name: string,
+    age: number,
+    email: string,
+    address: Address
+}
