+++ week-1-ts-ecosystem/w01e01-typescript-basics/src/User.ts
@@ -1,0 +1,11 @@
+import { UserType } from './User.d'
+
+export const user: UserType = {
+    name: 'Andy',
+    age: 30,
+    email: 'andy@mail-me-tommorow.com',
+    address: {
+        street: 'Strange Alley',
+        no: 23,
+    },
+}
